requirements.txt
python-dotenv==0.21.0
requests==2.28.2
psycopg2-binary==2.9.5